### 2024-09-10 #656 @aourednik
SPARQLChunks : query SPARQL endpoints and run SPARQL chunks in R-markdown files.
Submitting Author Name: André Ourendik
Submitting Author Github Handle: @aourednik
Repository:  https://github.com/aourednik/SPARQLchunks
Version submitted:0.2.1
Submission type: Standard

Version accepted: TBD

---

```
Package: SPARQLchunks
Title: Run SPARQL Chunks and Inline Functions to Retrieve Data
Version: 0.3.0
    person("Andre", "Ourednik", , "andre.ourednik@epfl.ch", role = c("aut", "cre"),
Description: Allows to run SPARQL chunks in R-markdown files. Also provides inline functions to send a SPARQL query to an endpoint and retrieve data in data frame or list form.
    curl,
    httr,
    xml2

```

SPARQL is a major language for accessing, extracting and munging RDF data. While RStudio provides a number of convenient ways to access to write SQL code chunks in Rmarkdown, to run these chunks and to assign the value of the query result directly to a variable of your choice. No such thing is available out of the box for SPARQL queries. This is where the SPARQLchunks package steps in. It allows you to query SPARQL endpoints in two different ways: (1) run SPARQL chunks in Rmarkdown files.
(2) run inline functions to send SPARQL queries to a user-defined endpoint and retrieve data in dataframe form (sparql2df) or list form (sparql2list). Endpoints can also be reached from behind corporate firewalls on Windows machines thanks to automatic proxy detection.

All scientists working with RDF data repositories

Not really, there used to be a "SPARQL" package, but it is deprecated. There is also the more genereic rdflib. No package except SPARQLChunks allows to write chunks of SPARQL in RMarkdown and retrieve the data directly into a data.frame or list object.

The package does not pass the automatic CRAN checks, as storing the result of a chunk requires assigning to the global
environment to store the query result in the output.var.  Example:
