### 2024-10-15 #656 @adamhsparks
Hi @aourednik, have you made any progress addressing the issues that the bot flagged?

Please let me know how I can assist you in correcting any of these issues.
