### 2025-02-02 #656 @maurolepore
Dear @aourednik, I'm starting my EiC rotation.

I see a last contact intent by Emily more than one month ago (on Nov 10) and no response after that. In line with our policy (pasted below) I'll close this issue and also let you know also by email.

I totally understand there can be many reasons why you're unable to respond. If you're ever able to continue please do not hesitate to resubmit.

--

https://devguide.ropensci.org/softwarereview_policies.html#policiesreviewprocess

> * If the author hasn’t requested a holding label, but is simply not responding, we should close the issue within one month after the last contact intent. This intent will include a comment tagging the author, but also an email using the email address listed in the DESCRIPTION of the package which is one of the rare cases where the editor will try to contact the author by email.
> * If a submission is closed and the author wishes to re-submit, they’ll have to start a new submission. If the package is still in scope, the author will have to respond to the initial reviews before the editor starts looking for new reviewers.
