### 2025-06-25 #656 @mpadge
Ah, I think that the issue needs to be open for the bot to respond. Responses to your concerns:

- The check for continuous integration is indeed confusing, and shall be addressed in https://github.com/ropensci-review-tools/pkgcheck/issues/255 to say in your case the checks are present, yet fail. Sorry about the confusion there.
- The note about not having examples is because you need to add `#' @noRd` to the Roxygen blocks to prevent the `.Rd` files from being generated in the `man/` directory. See https://devguide.ropensci.org/pkg_building.html#roxygen-2-use for details.
- The note about no return values is because the Roxygen2 tag is `@return` and not `@returns`.

More generally, please consider that these checks and associated reports serve the singular purpose of helping you to get your package in shape, rather than being "counterfactual" in any way. There is a wealth of information in https://devguide.ropensci.org, and all the links included there. Note also that I can see workflows on your repo when I use the GitHub API, and yet I see no "Actions" tab at the top, which should be present once workflows have been triggered? If you use [`usethis::use_github_action()`](https://usethis.r-lib.org/reference/use_github_action.html), it should definitely ensure that everything is set up correctly.

Hope that helps
