### 2025-06-26 #656 @aourednik
Many thanks @mpadge , this helps a lot. I've changed all the points you mention in my code. On my local machine, all unit tests I'v designed pass, and I have 79% coverage. But they fail on GitHub. Perhaps because the package is web dependant and all my functions call an external resource with httr ?
