### 2025-07-05 #656 @adamhsparks
Hi @aourednik, I'll be handling your submission as your editor. Looking at the package checks, I see it's only failing on Windows, https://github.com/aourednik/SPARQLchunks/actions/runs/15945383670/job/44978766469.

Is this a real issue or is it just a CI hiccup? Could you re-run the failing check and see if it passes on Windows or correct the issue so that we can get started on the review?
